skills/wechat-mp-typeset/scripts/requirements.txt
markdown>=3.4.0
PyYAML>=6.0.0
